[functional tests] Modify expected result after moving files
