+++ .pre-commit-config.yaml
@@ -4,7 +4,7 @@ repos:
     hooks:
     - id: black
       args: [--safe, --quiet]
-      exclude: tests/functional|tests/input|tests/extensions/data|tests/regrtest_data/|tests/data/
+      exclude: tests/functional/|tests/input|tests/extensions/data|tests/regrtest_data/|tests/data/
 -   repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v2.1.0
     hooks:
