+++ tests/functional/i/import_error.txt
@@ -1,4 +1,4 @@
 import-error:3::"Unable to import 'totally_missing'"
 import-error:16::"Unable to import 'maybe_missing_2'"
-import-error:27::"Unable to import 'functional.collections'"
+import-error:27::"Unable to import 'functional.c.collections'"
 syntax-error:29::"Cannot import 'syntax_error' due to syntax error 'invalid syntax (<unknown>, line 1)'"
