+++ tests/functional/m/method_hidden.txt
@@ -1,1 +1,1 @@
-method-hidden:13:Cdef.abcd:An attribute defined in functional.method_hidden line 9 hides this method
+method-hidden:13:Cdef.abcd:An attribute defined in functional.m.method_hidden line 9 hides this method
