+++ tests/functional/n/no_name_in_module.txt
@@ -10,5 +10,5 @@ no-name-in-module:34::No name 'anything' in module 'collections'
 no-name-in-module:44::No name 'indeed_missing' in module 'collections'
 no-name-in-module:49::No name 'emit' in module 'collections'
 no-name-in-module:60::No name 'emit2' in module 'collections'
-no-name-in-module:65::No name 'lala' in module 'functional.no_self_use'
-no-name-in-module:66::No name 'bla' in module 'functional.no_self_use'
+no-name-in-module:65::No name 'lala' in module 'functional.n.no_self_use'
+no-name-in-module:66::No name 'bla' in module 'functional.n.no_self_use'
